# Only copy bytes when required, otherwise use ranges of the span