@Henr1k80 2025-05-01
> let me know if you think this works better

Mostly looks prettier, but performance is lower?

Can you confirm the performance difference on a desktop?

| Method                                 | Mean        | Error     | StdDev    | Gen0   | Gen1   | Allocated |
|--------------------------------------- |------------:|----------:|----------:|-------:|-------:|----------:|
| DotNet_Base64                          |   128.69 ns |  2.615 ns |  5.401 ns | 0.0138 |      - |      88 B |
| SimpleBase_Base16_UpperCase            |    76.96 ns |  1.595 ns |  3.943 ns | 0.0101 |      - |      64 B |
| SimpleBase_Base16_UpperCase_TextReader |   475.95 ns |  9.094 ns |  9.339 ns | 1.3347 | 0.0005 |    8376 B |
| SimpleBase_Base32_Crockford            |   215.56 ns |  4.771 ns | 14.066 ns | 0.0126 |      - |      80 B |
| SimpleBase_Base85_Z85                  |   333.31 ns |  6.781 ns | 11.698 ns | 0.0138 |      - |      88 B |
| SimpleBase_Base58_Bitcoin              | 4,558.55 ns | 81.499 ns | 72.247 ns | 0.0076 |      - |      88 B |
| SimpleBase_Base58_Monero               |   145.96 ns |  2.995 ns |  3.205 ns | 0.0138 |      - |      88 B |