@ssg 2025-05-01
> Mostly looks prettier, but performance is lower?
> 
> Can you confirm the performance difference on a desktop?

they look identical on my machine (actually slightly slower without Range):

with Range:

BenchmarkDotNet v0.14.0, Windows 11 (10.0.26100.3915)
AMD Ryzen 9 5950X, 1 CPU, 32 logical and 16 physical cores
.NET SDK 8.0.408
  [Host]     : .NET 8.0.15 (8.0.1525.16413), X64 RyuJIT AVX2
  DefaultJob : .NET 8.0.15 (8.0.1525.16413), X64 RyuJIT AVX2


| Method                                 | Mean        | Error     | StdDev    | Gen0   | Gen1   | Allocated |
|--------------------------------------- |------------:|----------:|----------:|-------:|-------:|----------:|
| DotNet_Base64                          |   103.01 ns |  0.639 ns |  0.598 ns | 0.0052 |      - |      88 B |
| SimpleBase_Base16_UpperCase            |    50.30 ns |  0.340 ns |  0.318 ns | 0.0038 |      - |      64 B |
| SimpleBase_Base16_UpperCase_TextReader |   273.77 ns |  5.506 ns | 10.206 ns | 0.5007 | 0.0153 |    8376 B |
| SimpleBase_Base32_Crockford            |   128.11 ns |  1.435 ns |  1.342 ns | 0.0048 |      - |      80 B |
| SimpleBase_Base85_Z85                  |   250.24 ns |  1.603 ns |  1.499 ns | 0.0052 |      - |      88 B |
| SimpleBase_Base58_Bitcoin              | 3,679.22 ns | 20.047 ns | 17.771 ns | 0.0038 |      - |      88 B |
| SimpleBase_Base58_Monero               |   128.54 ns |  1.521 ns |  1.349 ns | 0.0052 |      - |      88 B |

without Range:

BenchmarkDotNet v0.14.0, Windows 11 (10.0.26100.3915)
AMD Ryzen 9 5950X, 1 CPU, 32 logical and 16 physical cores
.NET SDK 8.0.408
  [Host]     : .NET 8.0.15 (8.0.1525.16413), X64 RyuJIT AVX2
  DefaultJob : .NET 8.0.15 (8.0.1525.16413), X64 RyuJIT AVX2


| Method                                 | Mean        | Error     | StdDev    | Gen0   | Gen1   | Allocated |
|--------------------------------------- |------------:|----------:|----------:|-------:|-------:|----------:|
| DotNet_Base64                          |   104.13 ns |  1.869 ns |  1.657 ns | 0.0052 |      - |      88 B |
| SimpleBase_Base16_UpperCase            |    50.42 ns |  0.775 ns |  0.647 ns | 0.0038 |      - |      64 B |
| SimpleBase_Base16_UpperCase_TextReader |   266.10 ns |  5.332 ns | 14.323 ns | 0.5007 | 0.0153 |    8376 B |
| SimpleBase_Base32_Crockford            |   128.75 ns |  2.637 ns |  2.708 ns | 0.0048 |      - |      80 B |
| SimpleBase_Base85_Z85                  |   253.05 ns |  2.805 ns |  3.001 ns | 0.0052 |      - |      88 B |
| SimpleBase_Base58_Bitcoin              | 3,704.43 ns | 27.532 ns | 25.753 ns | 0.0038 |      - |      88 B |
| SimpleBase_Base58_Monero               |   129.17 ns |  2.127 ns |  1.776 ns | 0.0052 |      - |      88 B |