@ssg 2025-05-01
oh, I also noticed that this was a breaking change for `Decode()`. We can't merge it to this branch, but let me review it more carefully for 5.x.