@ssg 2025-05-01
Ah, okay. `Decode()` doesn't change. only `internalDecode()` does. Then it's not a breaking change. Let me go ahead with the merge if you're happy with the benchmark results.