@ssg 2025-05-02
Okay, merging this. I think [PublicApi.Analyzer](https://www.nuget.org/packages/Microsoft.CodeAnalysis.PublicApiAnalyzers/) does something like that, but I was just put off by GitHub's diff compression: it showed `Decode()` function declaration at the top of the first code change :)  Thanks!